requirements.txt
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.1
requests==2.26.0
werkzeug==2.0.1
pymysql==1.0.2
SQLAlchemy==1.4.23
